Specialise ARM libcxx buildbot to work around bugs
There are some inconsistencies in the way linker flags are passed to the lit tests,
as well as problems in the Clang driver when selecting compiler-rt paths and libraries,
as well as conflicts between compiler-rt+libcxxabi and libgcc_s/libgcc_eh.

While it could take months to fix them all, we need the bot running as soon as possible,
so working around all those problems at once seem only logical.

There's one missing failure that needs further investigation, as it doens't seem like a
build issue, but a real bug.